test: Move `side_effects` and `FakeMessage` test utilities to common file
This conflates "a file with tests for utilities" and "a file with
utilities for tests" but if you're OK with that, I'm OK with that.
